+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `ArrayExtensions` to be more specific
   - Class based operations `FirstIndexOfItem()` and `LastIndexOfItem()`
   - Struct based operations `FirstIndexOfValue()` and `LastIndexOfValue()`
+  - Correctly moved into GDX namespace.
 - Changed `ListExtensions` to be more specific.
   - Class based operations `ContainsItem()`, `RemoveItems()`, `RemoveFirstItem()` and `RemoveLastItem()`..
   - Struct based operations `ContainsValue()`, `RemoveValues()`, `RemoveFirstValue()` and `RemoveLastValue()`.